deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy
scikit-learn==1.4.1.post1
Pillow
plotly